+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/exceptions/ModuleException.java
@@ -1,0 +1,10 @@
+package nl.jqno.equalsverifier.internal.exceptions;
+
+/** Signals that a class was inaccessible. */
+@SuppressWarnings("serial")
+public class ModuleException extends MessagingException {
+
+    public ModuleException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}